Add org HTML export command for data URI
So I just want to export my org section to HTML in order to open it in
a graphical browser, and copy it to clipboard as rich text.  Normally
you'd export to a file, transfer the file to wherever the graphical
browser is running, and open the file.  Let's cut out a few steps by
directly opening a data URI containing the exported HTML.